Moved addresses from Store reducer to Delivery reducer
