+++ src/redux/Delivery/reducers.js
@@ -1,11 +1,16 @@
+import _ from 'lodash';
+
 import {
   ASSERT_DELIVERY_ERROR,
+  LOAD_ADDRESSES_SUCCESS,
   SET_RETURN_SCREEN
 } from "./actions"
 
+
 const initialState = {
-  returnScreen: '',
+  addresses: [],
   assertDeliveryError: null,
+  returnScreen: '',
 }
 
 export default (state = initialState, action = {}) => {
@@ -21,6 +26,16 @@ export default (state = initialState, action = {}) => {
         ...state,
         assertDeliveryError: action.payload,
       };
+
+    case LOAD_ADDRESSES_SUCCESS:
+      if (action.payload.store['@id'] === state.store['@id']) {
+        return {
+          ...state,
+          addresses: _.uniqBy(action.payload.addresses, '@id'),
+        };
+      }
+
+      break;
   }
 
   return state;
